+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #96 DbTester - a JUnit 5 extension to bootstrap test databases, provide integration w/Testcontainers
 * #97 DbTester support for Liquibase 
 * #98 DbTester: support for custom init function
+* #99 DbTester: database reusable between test classes
 
 ## 1.1